Move ICON fix for all variables to base class
